Auto-unseal and irsa support for vault addon
There is a chicken or egg problem of needing the kms key info in order to create a custom vault config to pass to helm, but if the key is created outside of the module, and we try to base count off the existence of the arn string we end up with the "terraform cannot predict how many instances will be created" issue.  There is probably a better way to handle this, but I have not been able to piece it together.